Rework the repo to build edk2 release tags
The builder this repo used to pull from has been turned off. It would be
nice to have up-to-date edk2 release builds for use with testing in
https://github.com/rust-osdev/uefi-rs, without relying on the vagaries
of Linux distro releases. Doing our own builds also allows us to ensure
we have exactly the build options we want for use with VM testing.

As described in the readme, the intended usage is manually pushing a tag
corresponding to a release tag in the edk2 repo. That tag will be built
in a container, and the resulting tarball will be pushed as a release in
this repo (as long as the release doesn't already exist.)

I've intentionally kept the container build very simple and linear to
make it clear what options are being set for each edk2 build. That leads
to some duplication there, but I think the end result is quite easy to
follow.